requirements - Copy.txt
openai
azure-search-documents
azure-ai-documentintelligence
python-dotenv
